begin working on the cable manager code
